testing/web-platform/tests/encoding-detection/th-windows-874.tentative.html

<!doctype html>
<title>th windows-874</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>����͡�÷��ͺ�����������ѡ���</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-874", 'Expected windows-874');
    }, "Check detection result");
      done();
};
</script>
